requirements.txt
kafka-python
black
isort